+++ tests/SEGDataOverlayNoHydrationThenMPR.spec.ts
@@ -10,19 +10,20 @@ test.beforeEach(async ({ page }) => {
 test('should launch MPR with unhydrated SEG chosen from the data overlay menu', async ({
   page,
   mainToolbarPageObject,
+  rightPanelPageObject,
   viewportPageObject,
 }) => {
-  await page.getByTestId('side-panel-header-right').click();
-  await viewportPageObject.getById('default').overlayMenu.dataOverlay.click();
-  await page.getByTestId('AddSegmentationDataOverlay-default').click();
-  await page.getByText('SELECT A SEGMENTATION').click();
-  await page.getByTestId('Segmentation').click();
+  await rightPanelPageObject.toggle();
+  await viewportPageObject.getById('default').overlayMenu.dataOverlay.toggle();
+  await viewportPageObject
+    .getById('default')
+    .overlayMenu.dataOverlay.addSegmentation('Segmentation');
 
   // Adding an overlay should not show the LOAD button.
   assertNumberOfModalityLoadBadges({ page, expectedCount: 0 });
 
   // Hide the overlay menu.
-  await viewportPageObject.getById('default').overlayMenu.dataOverlay.click();
+  await viewportPageObject.getById('default').overlayMenu.dataOverlay.toggle();
 
   await page.waitForTimeout(5000);